# src/modules/api/openapi/spec.yaml
openapi: 3.0.3
info:
  title: Salina ERP API
  version: "1.0.0"
  description: |
    REST API for Salina ERP - Publisher Royalty Management Platform.

    ## Authentication
    All API requests require authentication via Bearer tokens. See the Authentication section for details.

    ## Rate Limits
    - Default: 100 requests/minute + 1000 requests/hour
    - Premium: 1000 requests/minute + 10000 requests/hour
    - Auth endpoint: 10 requests/minute (IP-based)

    Rate limit headers are included in all responses.
  contact:
    name: Salina Support
    email: support@salina.app
  license:
    name: Proprietary
    url: https://salina.app/terms
servers:
  - url: https://{tenant}.salina.app/api/v1
    description: Production API
    variables:
      tenant:
        default: demo
        description: Your tenant subdomain
...
security:
  - bearerAuth: []
paths:
...
  /titles/{id}:
    get:
      tags: [Titles]
      summary: Get a title
      description: Retrieve a single title by ID.
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "200":
          description: Title details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitleResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Titles]
      summary: Update a title
      description: Update an existing title. Partial updates supported. Requires `write` scope.
      parameters:
        - $ref: "#/components/parameters/resourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TitleUpdate"
      responses:
        "200":
          description: Title updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitleResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /onix/export:
    get:
      tags: [ONIX]
      summary: Export ONIX metadata
      description: |
        Export title metadata in ONIX 3.1 (default) or 3.0 XML format.
        Use this endpoint to sync your catalog with distribution channels.
      parameters:
        - name: title_ids
          in: query
          description: Comma-separated list of title IDs to export
          schema:
            type: string
        - name: version
          in: query
          description: ONIX version
          schema:
            type: string
            enum: ["3.1", "3.0"]
            default: "3.1"
        - name: since
          in: query
          description: Only export titles updated after this timestamp
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: ONIX XML document
          content:
            application/xml:
              schema:
                type: string
                example: '<?xml version="1.0"?>...'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained via /auth/token endpoint.
        Include in Authorization header: `Bearer <token>`
  parameters:
...
    resourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: not_found
              message: Resource not found
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: validation_error
              message: Validation failed
              details:
                title: Title is required
                isbn: Invalid ISBN-13 format
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [validation_error, not_found, unauthorized, forbidden, rate_limited, server_error]
            message:
              type: string
            details:
              type: object
              additionalProperties:
                type: string
...
    Title:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        subtitle:
          type: string
        isbn:
          type: string
          description: ISBN-13
        asin:
          type: string
          description: Amazon Standard Identification Number
        format:
          type: string
          enum: [paperback, hardcover, ebook, audiobook]
        publication_status:
          type: string
          enum: [draft, pending, published, out_of_print]
        list_price:
          type: number
          format: decimal
        publication_date:
          type: string
          format: date
        authors:
          type: array
          items:
            type: object
            properties:
              contactId:
                type: string
                format: uuid
              name:
                type: string
              isPrimary:
                type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    TitleUpdate:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
        publication_status:
          type: string
          enum: [draft, pending, published, out_of_print]
        list_price:
          type: number
        publication_date:
          type: string
          format: date
        asin:
          type: string
    TitleResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Title"
...